assistant/requirements.txt
openai>=1.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
